pyproject.toml
[tool.poetry]
name = "chat-with-pdf"
version = "0.1.0"
description = "Chat with PDF Tutorial"
authors = ["Adrian Moses <adrian@adrianmoses.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.4"
python-dotenv = "^1.0.1"
PyPDF2 = "^3.0.1"
streamlit = "^1.30.0"
langchain = "^0.1.4"
openai = "^1.10.0"
chromadb = "^0.4.22"
langchain-community = "^0.0.16"
pypdf = "^4.0.1"
tiktoken = "^0.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

packages = [
    { include = "chat-with-pdf"}
]